move constants into feature folders
