+++ src/features/chart/redux/constants.js
@@ -1,0 +1,2 @@
+export const UPDATE_CHART_DATA = "UPDATE_CHART_DATA";
+export const UPDATE_CHART_RANGE = "UPDATE_CHART_RANGE";
