+++ src/features/company/redux/constants.js
@@ -1,0 +1,1 @@
+export const UPDATE_COMPANY_OVERVIEW = "UPDATE_COMPANY_OVERVIEW";
