+++ src/features/headline/redux/constants.js
@@ -1,0 +1,3 @@
+export const UPDATE_SEARCH = "UPDATE_SEARCH";
+export const UPDATE_COMPANY_SYMBOLS = "UPDATE_COMPANY_SYMBOLS ";
+export const UPDATE_SEARCH_QUERY = "UPDATE_SEARCH_QUERY";
