+++ src/features/key-stats/redux/actions.js
@@ -1,4 +1,4 @@
-import { UPDATE_KEY_STATS } from "../../../store/constants";
+import { UPDATE_KEY_STATS } from "./constants";
 
 export const updateKeyStatsAction = action => ({
   type: UPDATE_KEY_STATS,
