+++ src/features/key-stats/redux/constants.js
@@ -1,0 +1,1 @@
+export const UPDATE_KEY_STATS = "UPDATE_KEY_STATS";
