+++ src/features/latest-news/redux/actions.js
@@ -1,4 +1,4 @@
-import { UPDATE_LATEST_NEWS } from "../../../store/constants";
+import { UPDATE_LATEST_NEWS } from "./constants";
 
 export const updateLatestNewsAction = action => ({
   type: UPDATE_LATEST_NEWS,
