+++ src/features/latest-news/redux/constants.js
@@ -1,0 +1,1 @@
+export const UPDATE_LATEST_NEWS = "UPDATE_LATEST_NEWS";
