+++ src/features/peers/redux/actions.js
@@ -1,4 +1,4 @@
-import { UPDATE_TOP_PEERS } from "../../../store/constants";
+import { UPDATE_TOP_PEERS } from "./constants";
 
 export const updateTopPeersAction = action => ({
   type: UPDATE_TOP_PEERS,
