+++ src/features/peers/redux/constants.js
@@ -1,0 +1,1 @@
+export const UPDATE_TOP_PEERS = "UPDATE_TOP_PEERS";
