+++ src/features/peers/redux/reducer.js
@@ -1,4 +1,5 @@
-import { UPDATE_TOP_PEERS, RESET } from "../../../store/constants";
+import { RESET } from "../../../store/constants";
+import { UPDATE_TOP_PEERS } from "./constants";
 
 const initialState = { selectedTopPeers: null };
 export const peerReducer = (state = initialState, { type, payload }) => {
