+++ src/features/search/index.js
@@ -1,2 +1,1 @@
 export { Search } from "./components/Search";
-export { searchMiddleware } from "./redux/middleware";
