+++ src/features/stock-ticker/redux/actions.js
@@ -1,4 +1,4 @@
-import { UPDATE_STOCK_TICKER } from "../../../store/constants";
+import { UPDATE_STOCK_TICKER } from "./constants";
 
 export const updateStockTickerAction = action => ({
   type: UPDATE_STOCK_TICKER,
