+++ src/features/stock-ticker/redux/constants.js
@@ -1,0 +1,2 @@
+export const UPDATE_STOCK_TICKER = "UPDATE_STOCK_TICKER";
+export const UPDATE_SELECTED_STOCK = "UPDATE_SELECTED_STOCK";
