+++ src/store/constants.js
@@ -1,14 +1,3 @@
-export const ON_STOCK_DATA_RECIEVED = "ON_STOCK_DATA_RECIEVED";
-export const UPDATE_SELECTED_STOCK = "UPDATE_SELECTED_STOCK";
-export const UPDATE_CHART_RANGE = "UPDATE_CHART_RANGE";
 export const BOOTSTRAP = "BOOTSTRAP";
-export const UPDATE_KEY_STATS = "UPDATE_KEY_STATS";
-export const UPDATE_LATEST_NEWS = "UPDATE_LATEST_NEWS";
-export const UPDATE_COMPANY_OVERVIEW = "UPDATE_COMPANY_OVERVIEW";
-export const UPDATE_TOP_PEERS = "UPDATE_TOP_PEERS";
-export const UPDATE_SEARCH = "UPDATE_SEARCH";
-export const UPDATE_COMPANY_SYMBOLS = "UPDATE_COMPANY_SYMBOLS ";
-export const UPDATE_STOCK_TICKER = "UPDATE_STOCK_TICKER";
-export const UPDATE_CHART_DATA = "UPDATE_CHART_DATA";
-export const UPDATE_SEARCH_QUERY = "UPDATE_SEARCH_QUERY";
 export const RESET = "RESET";
+export const UPDATE_SELECTED_STOCK = "UPDATE_SELECTED_STOCK";
