+++ src/store/store.js
@@ -1,6 +1,5 @@
 import { createStore, applyMiddleware, compose } from "redux";
 import { stockMiddleware } from "../features/stock-ticker";
-import { searchMiddleware } from "../features/search";
 import { chartMiddleware } from "../features/chart";
 import { combinedReducer } from "./reducer";
 import { socketService } from "../services";
@@ -14,7 +13,6 @@ const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const middleware = [
   stockMiddleware({ socketService }),
   chartMiddleware({ socketService }),
-  searchMiddleware({ socketService }),
   companyMiddleware({ socketService }),
   keyStatsMiddleware({ socketService }),
   latestNewsMiddleware({ socketService }),
